HSEARCH-4203 Move Deadline to API
