+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/loading/impl/PojoMultiLoaderLoadingPlan.java
@@ -9,7 +9,7 @@ package org.hibernate.search.mapper.pojo.loading.impl;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-import org.hibernate.search.engine.common.timing.spi.Deadline;
+import org.hibernate.search.engine.common.timing.Deadline;
 import org.hibernate.search.mapper.pojo.loading.spi.PojoSelectionLoadingContext;
 import org.hibernate.search.mapper.pojo.loading.spi.PojoLoadingTypeContext;
 import org.hibernate.search.mapper.pojo.loading.spi.PojoSelectionLoadingStrategy;
