+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/search/loading/impl/PojoProjectionHitMapper.java
@@ -11,7 +11,7 @@ import java.util.Map;
 
 import org.hibernate.search.engine.backend.common.DocumentReference;
 import org.hibernate.search.engine.backend.common.spi.DocumentReferenceConverter;
-import org.hibernate.search.engine.common.timing.spi.Deadline;
+import org.hibernate.search.engine.common.timing.Deadline;
 import org.hibernate.search.engine.search.loading.spi.LoadingResult;
 import org.hibernate.search.engine.search.loading.spi.ProjectionHitMapper;
 import org.hibernate.search.mapper.pojo.bridge.runtime.spi.BridgeSessionContext;
